backend/requirements.txt
alpha-vantage==2.1.1
alphavantage==0.0.11
aniso8601==6.0.0
astroid==2.3.1
awscli==1.16.257
blis==0.2.4
boto3==1.9.224
botocore==1.12.247
certifi==2019.9.11
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
Click==7.0
colorama==0.3.9
cycler==0.10.0
cymem==2.0.2
Django==2.2.4
django-admin==1.3.2
django-cors-headers==3.1.0
django-excel-response2==2.0.8
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.10.2
dnspython==1.16.0
docutils==0.15.2
elasticsearch==7.0.5
Flask==1.1.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
Flask-S3==0.3.3
Flask-SQLAlchemy==2.4.0
foil==0.2.7
freegames==2.3.1
gcloud==0.17.0
googleapis-common-protos==1.6.0
googlemaps==3.1.1
graphene==2.1.7
graphene-django==2.4.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
httplib2==0.13.1
ibm-cloud-sdk-core==1.0.0
ibm-watson==4.0.1
idna==2.8
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
jsonify==0.5
jws==0.1.3
kaggle==1.5.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
lxml==4.4.1
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.1.1
mccabe==0.6.1
multitasking==0.0.9
murmurhash==1.0.2
nltk==3.4.5
nodeenv==1.3.3
numpy==1.17.3
oauth2client==3.0.0
pandas==0.25.1
pandas-datareader==0.8.1
pandas-market-calendars==1.1
ParallelDots==3.2.13
Pint==0.9
pipenv==2018.11.26
plac==0.9.6
preshed==2.0.1
promise==2.2.1
protobuf==3.10.0rc1
psycopg2==2.8.3
psycopg2-binary==2.8.3
py-bcrypt==0.4
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycryptodome==3.4.3
PyJWT==1.7.1
pylint==2.4.2
pymongo==3.9.0
PyMySQL==0.9.3
pyparsing==2.4.2
Pyrebase==3.0.27
PySocks==1.7.0
python-dateutil==2.7.5
python-http-client==3.1.0
python-jwt==2.0.1
python-slugify==3.0.6
pytz==2019.3
PyYAML==3.13
requests==2.22.0
requests-aws4auth==0.9
requests-toolbelt==0.7.0
rsa==3.4.2
Rx==1.6.1
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
screen==1.0.1
sendgrid==6.0.5
singledispatch==3.4.0.3
six==1.12.0
spacy==2.1.9
SQLAlchemy==1.2.19
sqlparse==0.3.0
srsly==0.2.0
text-unidecode==1.3
thinc==7.0.8
tqdm==4.37.0
twilio==6.30.0
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.6.0
virtualenv-clone==0.5.3
wasabi==0.3.0
websocket-client==0.48.0
Werkzeug==0.16.0
whitenoise==3.3.1
wrapt==1.11.2
xlwt==1.3.0
yfinance==0.1.45
